deleting added, show movie added, user menu added

+++ src/main/java/com/tms/stankevich/service/MovieServiceImpl.java
@@ -4,6 +4,9 @@ import com.tms.stankevich.dao.GenreRepository;
 import com.tms.stankevich.dao.MovieRepository;
 import com.tms.stankevich.domain.movie.Genre;
 import com.tms.stankevich.domain.movie.Movie;
+import com.tms.stankevich.domain.movie.Session;
+import com.tms.stankevich.exception.GenreDeleteException;
+import com.tms.stankevich.exception.MovieDeleteException;
 import com.tms.stankevich.service.MovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,6 +22,9 @@ public class MovieServiceImpl implements MovieService {
 
     @Autowired
     private GenreRepository genreRepository;
+
+    @Autowired
+    private SessionService sessionService;
 
     @Override
     public List<Movie> getAllMovies() {
@@ -51,12 +57,18 @@ public class MovieServiceImpl implements MovieService {
     }
 
     @Override
-    public void deleteGenre(Genre genre) {
+    public void deleteGenre(Genre genre) throws GenreDeleteException {
+        List<Movie> movies = movieRepository.findByGenre(genre);
+        if (movies.size() > 0)
+            throw new GenreDeleteException(movies.size());
         genreRepository.delete(genre);
     }
 
     @Override
-    public void deleteMovie(Movie movie) {
+    public void deleteMovie(Movie movie) throws MovieDeleteException {
+        List<Session> sessions = sessionService.findByMovie(movie);
+        if (sessions.size() > 0)
+            throw new MovieDeleteException(sessions.size());
         movieRepository.delete(movie);
     }
 }
